+++ features/support/env.rb
@@ -1,4 +1,3 @@
-# require 'mini/test'
 
 $LOAD_PATH.unshift(File.dirname(__FILE__) + '/../../lib')
 require 'bio-vcf'
@@ -7,7 +6,7 @@ require 'rspec/expectations'
 
 # Add the regression module if in the path (it can also be a gem)
 rootdir = File.dirname(__FILE__) + '/../..'
-$LOAD_PATH.unshift(rootdir+'/lib',rootdir+'/../regressiontest/lib')
+$LOAD_PATH.unshift(rootdir+'/lib/regressiontest',rootdir+'/../regressiontest/lib')
 require 'regressiontest'
 
 include BioVcf